Export command deprecation.
