+++ commands/util.go
@@ -31,7 +31,6 @@ func NewRest() *rest.Rest {
 		call.NewUserCalls(),
 		call.NewProjectCalls(),
 		call.NewContractCalls(),
-		call.NewExportCalls(),
 		call.NewNetworkCalls(),
 		call.NewActionCalls(),
 		call.NewDevNetCalls(),
