+++ config/config.go
@@ -3,13 +3,10 @@ package config
 import (
 	"flag"
 	"fmt"
-	"math/big"
 	"os"
 	"os/user"
 	"path/filepath"
 
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/params"
 	"github.com/tenderly/tenderly-cli/model/actions"
 	extensionsModel "github.com/tenderly/tenderly-cli/model/extensions"
 	"github.com/tenderly/tenderly-cli/userError"
@@ -30,7 +27,6 @@ const (
 
 	OrganizationName = "org_name"
 
-	Exports    = "exports"
 	Actions    = "actions"
 	Extensions = "node_extensions"
 	Projects   = "projects"
@@ -38,147 +34,6 @@ const (
 
 var defaultsGlobal = map[string]interface{}{
 	Token: "",
-}
-
-type EthashConfig struct{}
-
-type CliqueConfig struct {
-	Period uint64 `mapstructure:"period"`
-	Epoch  uint64 `mapstructure:"epoch"`
-}
-
-type BigInt interface{}
-
-func toInt(x BigInt) (*big.Int, error) {
-	if x == nil {
-		return nil, nil
-	}
-
-	if stringVal, ok := x.(string); ok {
-		i := &big.Int{}
-		_, ok := i.SetString(stringVal, 10)
-		if !ok {
-			return nil, fmt.Errorf("failed parsing big int: %s", stringVal)
-		}
-
-		return i, nil
-	}
-
-	if numberVal, ok := x.(int64); ok {
-		return big.NewInt(numberVal), nil
-	}
-
-	if numberVal, ok := x.(int); ok {
-		return big.NewInt(int64(numberVal)), nil
-	}
-
-	return nil, fmt.Errorf("unrecognized value: %s", x)
-}
-
-type ChainConfig struct {
-	HomesteadBlock BigInt `mapstructure:"homestead_block,omitempty" yaml:"homestead_block,omitempty"`
-
-	EIP150Block BigInt      `mapstructure:"eip150_block,omitempty" yaml:"eip150_block,omitempty"`
-	EIP150Hash  common.Hash `mapstructure:"eip150_hash,omitempty" yaml:"eip150_hash,omitempty"`
-
-	EIP155Block BigInt `mapstructure:"eip155_block,omitempty" yaml:"eip155_block,omitempty"`
-	EIP158Block BigInt `mapstructure:"eip158_block,omitempty" yaml:"eip158_block,omitempty"`
-
-	ByzantiumBlock      BigInt `mapstructure:"byzantium_block,omitempty" yaml:"byzantium_block,omitempty"`
-	ConstantinopleBlock BigInt `mapstructure:"constantinople_block,omitempty" yaml:"constantinople_block,omitempty"`
-	PetersburgBlock     BigInt `mapstructure:"petersburg_block,omitempty" yaml:"petersburg_block,omitempty"`
-	IstanbulBlock       BigInt `mapstructure:"istanbul_block,omitempty" yaml:"istanbul_block,omitempty"`
-	BerlinBlock         BigInt `mapstructure:"berlin_block,omitempty" yaml:"berlin_block,omitempty"`
-	LondonBlock         BigInt `mapstructure:"london_block,omitempty" yaml:"london_block,omitempty"`
-}
-
-var DefaultChainConfig = &ChainConfig{
-	HomesteadBlock:      0,
-	EIP150Block:         0,
-	EIP150Hash:          common.Hash{},
-	EIP155Block:         0,
-	EIP158Block:         0,
-	ByzantiumBlock:      0,
-	ConstantinopleBlock: 0,
-	PetersburgBlock:     0,
-	IstanbulBlock:       0,
-	BerlinBlock:         0,
-	LondonBlock:         0,
-}
-
-func (c *ChainConfig) Config() (*params.ChainConfig, error) {
-	homesteadBlock, err := toInt(c.HomesteadBlock)
-	if err != nil {
-		return nil, err
-	}
-
-	eip150Block, err := toInt(c.EIP150Block)
-	if err != nil {
-		return nil, err
-	}
-
-	eip155Block, err := toInt(c.EIP155Block)
-	if err != nil {
-		return nil, err
-	}
-
-	eip158Block, err := toInt(c.EIP158Block)
-	if err != nil {
-		return nil, err
-	}
-
-	byzantiumBlock, err := toInt(c.ByzantiumBlock)
-	if err != nil {
-		return nil, err
-	}
-
-	constantinopleBlock, err := toInt(c.ConstantinopleBlock)
-	if err != nil {
-		return nil, err
-	}
-
-	petersburgBlock, err := toInt(c.PetersburgBlock)
-	if err != nil {
-		return nil, err
-	}
-
-	istanbulBlock, err := toInt(c.IstanbulBlock)
-	if err != nil {
-		return nil, err
-	}
-
-	berlinBlock, err := toInt(c.BerlinBlock)
-	if err != nil {
-		return nil, err
-	}
-
-	londonBlock, err := toInt(c.LondonBlock)
-	if err != nil {
-		return nil, err
-	}
-
-	return &params.ChainConfig{
-		HomesteadBlock:      homesteadBlock,
-		EIP150Block:         eip150Block,
-		EIP150Hash:          c.EIP150Hash,
-		EIP155Block:         eip155Block,
-		EIP158Block:         eip158Block,
-		ByzantiumBlock:      byzantiumBlock,
-		ConstantinopleBlock: constantinopleBlock,
-		PetersburgBlock:     petersburgBlock,
-		IstanbulBlock:       istanbulBlock,
-		BerlinBlock:         berlinBlock,
-		LondonBlock:         londonBlock,
-	}, nil
-}
-
-type ExportNetwork struct {
-	Name          string              `mapstructure:"-"`
-	ProjectSlug   string              `mapstructure:"project_slug"`
-	RpcAddress    string              `mapstructure:"rpc_address"`
-	Protocol      string              `mapstructure:"protocol"`
-	ForkedNetwork string              `mapstructure:"forked_network"`
-	ChainConfig   *params.ChainConfig `mapstructure:"chain_config"`
 }
 
 var defaultsProject = map[string]interface{}{
@@ -238,11 +93,6 @@ func Init() {
 	}
 }
 
-func GetBool(key string) bool {
-	check(key)
-	return getBool(key)
-}
-
 func GetString(key string) string {
 	check(key)
 	return getString(key)
@@ -291,52 +141,6 @@ func IsLoggedIn() bool {
 
 func IsProjectInit() bool {
 	return getString(ProjectSlug) != "" || len(MaybeGetMap(Projects)) > 0
-}
-
-func IsNetworkConfigured(network string) bool {
-	if _, ok := getStringMapString(Exports)[network]; ok {
-		return true
-	}
-
-	return false
-}
-
-func WriteExportNetwork(networkId string, network *ExportNetwork) error {
-	exports := projectConfig.GetStringMap(Exports)
-
-	chainConfig := DefaultChainConfig
-	if network.ChainConfig != nil {
-		chainConfig = &ChainConfig{
-			HomesteadBlock:      network.ChainConfig.HomesteadBlock,
-			EIP150Block:         network.ChainConfig.EIP150Block,
-			EIP150Hash:          network.ChainConfig.EIP150Hash,
-			EIP155Block:         network.ChainConfig.EIP158Block,
-			EIP158Block:         network.ChainConfig.EIP158Block,
-			ByzantiumBlock:      network.ChainConfig.ByzantiumBlock,
-			ConstantinopleBlock: network.ChainConfig.ConstantinopleBlock,
-			PetersburgBlock:     network.ChainConfig.PetersburgBlock,
-			IstanbulBlock:       network.ChainConfig.IstanbulBlock,
-			BerlinBlock:         network.ChainConfig.BerlinBlock,
-			LondonBlock:         network.ChainConfig.LondonBlock,
-		}
-	}
-
-	exports[networkId] = struct {
-		ProjectSlug   string       `mapstructure:"project_slug" yaml:"project_slug"`
-		RpcAddress    string       `mapstructure:"rpc_address" yaml:"rpc_address"`
-		Protocol      string       `mapstructure:"protocol" yaml:"protocol"`
-		ForkedNetwork string       `mapstructure:"forked_network" yaml:"forked_network"`
-		ChainConfig   *ChainConfig `mapstructure:"chain_config" yaml:"chain_config"`
-	}{
-		ProjectSlug:   network.ProjectSlug,
-		RpcAddress:    network.RpcAddress,
-		Protocol:      network.Protocol,
-		ForkedNetwork: network.ForkedNetwork,
-		ChainConfig:   chainConfig,
-	}
-
-	projectConfig.Set(Exports, exports)
-	return WriteProjectConfig()
 }
 
 func IsAnyActionsInit() bool {
@@ -443,30 +247,6 @@ func getString(key string) string {
 	}
 
 	return globalConfig.GetString(key)
-}
-
-func getBool(key string) bool {
-	if projectConfig.IsSet(key) {
-		return projectConfig.GetBool(key)
-	}
-
-	return globalConfig.GetBool(key)
-}
-
-func getStringMapString(key string) map[string]interface{} {
-	if projectConfig.IsSet(key) {
-		return projectConfig.GetStringMap(key)
-	}
-
-	return globalConfig.GetStringMap(key)
-}
-
-func UnmarshalKey(key string, val interface{}) error {
-	if projectConfig.IsSet(key) {
-		return projectConfig.UnmarshalKey(key, val)
-	}
-
-	return globalConfig.UnmarshalKey(key, val)
 }
 
 func check(key string) {
