--- ethereum/README.md
@@ -1,3 +1,0 @@
-# Ethereum Client
-
-Ethereum client is supposed to be a general interface over any ethereum node.
