+++ main.go
@@ -10,7 +10,6 @@ import (
 	_ "github.com/tenderly/tenderly-cli/commands/actions"
 	_ "github.com/tenderly/tenderly-cli/commands/contract"
 	_ "github.com/tenderly/tenderly-cli/commands/devnet"
-	_ "github.com/tenderly/tenderly-cli/commands/export"
 	_ "github.com/tenderly/tenderly-cli/commands/extensions"
 )
